backend/pyproject.toml
[tool.poetry]
name = "fastapi-celery"
version = "0.1.0"
description = "A simple FastAPI server with Celery for background task processing"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.110.0"
celery = "^5.3.4"
redis = "^5.0.1"
uvicorn = "^0.24.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
email-validator = "^2.0.0"
httpx = ">=0.28.1"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
alembic = "^1.12.0"
requests = "^2.32.4"
google-genai = "^1.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
